refactor(wire): move AgentKind out of the CEF-linked crate
vmux_core links bevy_cef directly, so anything naming AgentKind is pulled
into the render stack. The headless service cannot pay that, and
AgentSession carries a kind.

vmux_wire is where it belongs anyway. Both avatar.rs and team.rs carried
doc comments apologising for the split — the agent palette is keyed by URL
segment because AgentKind was unreachable from the phone. It is reachable
now, so both notes go. The palette stays segment-keyed for the reason that
survives: a registry-driven ACP agent has no kind at all.

Keeps type_path = "vmux_core::agent" so reflect auto-registration and any
persisted store still resolve it. From<AgentKind> for TerminalKind stays in
vmux_core, which the orphan rule allows.
